refactored refl-cpp and (de-)serialiser to be an optional dependency
* fixed some implicit conversion warnings
* suppressed clang and gcc warnings for external dependencies

N.B. GR4 uses now it's own mini reflection library and yaml-(de-)serialiser -> perhaps this could be eventually upstreamed as well.
