+++ python/pmtv/bindings/pmt_python.cc
@@ -8,10 +8,9 @@
  */
 
 /* This file is automatically generated using bindtool */
-#ifdef __GNUC__
+#if defined(__clang__) || defined(__GNUC__)
 #pragma GCC diagnostic push // ignore warning of external libraries that from this lib-context we do not have any control over
 #pragma GCC diagnostic ignored "-Wall"
-#pragma GCC diagnostic ignored "-Wuseless-cast"
 #pragma GCC diagnostic ignored "-Wold-style-cast"
 #endif
 
@@ -23,7 +22,7 @@
 #define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
 #include <numpy/arrayobject.h>
 
-#ifdef __GNUC__
+#if defined(__clang__) || defined(__GNUC__)
 #pragma GCC diagnostic pop
 #endif
 
@@ -31,6 +30,7 @@ namespace py = pybind11;
 
 #include <pmtv/pmt.hpp>
 #include <pmtv/format.hpp>
+#include <pmtv/serialiser.hpp>
 
 // pydoc.h is automatically generated in the build directory
 // #include <pmt_pydoc.h>
@@ -61,7 +61,6 @@ py::object create_numpy_scalar(T val)
 #ifdef __GNUC__
 #pragma GCC diagnostic push // ignore warning of external libraries that from this lib-context we do not have any control over
 #pragma GCC diagnostic ignored "-Wall"
-#pragma GCC diagnostic ignored "-Wuseless-cast"
 #pragma GCC diagnostic ignored "-Wold-style-cast"
 #endif
     PyObject* scal =
