+++ python/pmtv/bindings/pmtv_python.cc
@@ -8,10 +8,13 @@
  *
  */
 
-#ifdef __GNUC__
+#if defined(__clang__) || defined(__GNUC__)
 #pragma GCC diagnostic push // ignore warning of external libraries that from this lib-context we do not have any control over
-#pragma GCC diagnostic ignored "-Wuseless-cast"
+#pragma GCC diagnostic ignored "-Wall"
 #pragma GCC diagnostic ignored "-Wold-style-cast"
+#ifndef __clang__ // only for GCC, not Clang
+#pragma GCC diagnostic ignored "-Wuseless-cast"
+#endif
 #endif
 
 #include <pybind11/pybind11.h>
@@ -19,7 +22,7 @@
 #define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
 #include <numpy/arrayobject.h>
 
-#ifdef __GNUC__
+#if defined(__clang__) || defined(__GNUC__)
 #pragma GCC diagnostic pop
 #endif
