+++ test/qa_vector_of_pmts.cpp
@@ -13,6 +13,7 @@
 
 #include <pmtv/pmt.hpp>
 #include <pmtv/format.hpp>
+#include <pmtv/serialiser.hpp>
 
 #include <fmt/core.h>
 
@@ -35,15 +36,14 @@ operator[] (lookup and add)
 Cheap copies (could be moves)
 */
 
-TEST(PmtVectorPmt, Constructor)
-{
+TEST(PmtVectorPmt, Constructor) {
     // Empty Constructor
-    pmt empty_vec{ std::vector<pmt>() };
+    pmt empty_vec{std::vector<pmt>()};
     EXPECT_EQ(std::get<std::vector<pmt>>(empty_vec).size(), 0);
-    pmt il_vec{ std::vector<pmt>{1.0, 2, "abc"}};
+    pmt il_vec{std::vector<pmt>{1.0, 2, "abc"}};
     std::vector<pmt> vec;
     vec.push_back(pmt(1));
-    vec.push_back(pmt(std::vector<uint32_t>{ 1, 2, 3 }));
+    vec.push_back(pmt(std::vector<uint32_t>{1, 2, 3}));
 
     auto p = pmt(vec);
 
@@ -53,11 +53,10 @@ TEST(PmtVectorPmt, Constructor)
     EXPECT_TRUE(vec[1] == vec2[1]);
 }
 
-TEST(PmtVectorPmt, fmt)
-{
+TEST(PmtVectorPmt, fmt) {
     std::vector<pmt> vec;
     vec.push_back(pmt(1));
-    vec.push_back(pmt(std::vector<uint32_t>{ 1, 2, 3 }));
+    vec.push_back(pmt(std::vector<uint32_t>{1, 2, 3}));
     EXPECT_EQ(fmt::format("{}", pmt(vec)), fmt::format("[{}]", fmt::join(vec, ", ")));
 
 }
